arch/x86: extract slaunch unit
To collect its core functionality in one place instead of having some in
intel_txt and other in tpm units.

TXT_EVTYPE_* now live in <asm/slaunch.h> and are called
DLE_EVTYPE_* despite being based on TXT specification.  This way code
for non-Intel won't need to include TXT header.

No functional changes.
