+++ xen/arch/x86/Makefile
@@ -59,6 +59,7 @@ obj-y += psr.o
 obj-y += intel_txt.o
 obj-y += setup.o
 obj-y += shutdown.o
+obj-y += slaunch.o
 obj-y += smp.o
 obj-y += smpboot.o
 obj-y += spec_ctrl.o
