+++ xen/arch/x86/cpu/common.c
@@ -15,7 +15,7 @@
 #include <asm/apic.h>
 #include <asm/random.h>
 #include <asm/setup.h>
-#include <asm/intel_txt.h>
+#include <asm/slaunch.h>
 #include <asm/shstk.h>
 #include <mach_apic.h>
 #include <public/sysctl.h> /* for XEN_INVALID_{SOCKET,CORE}_ID */
