+++ xen/arch/x86/setup.c
@@ -56,6 +56,7 @@
 #include <asm/microcode.h>
 #include <asm/pv/domain.h>
 #include <asm/intel_txt.h>
+#include <asm/slaunch.h>
 
 /* opt_nosmp: If true, secondary processors are ignored. */
 static bool __initdata opt_nosmp;
